+++ tests/e2e/e2e_evidence_test.go
@@ -8,8 +8,8 @@ import (
 	"cosmossdk.io/x/evidence/exported"
 	evidencetypes "cosmossdk.io/x/evidence/types"
 
-	"github.com/cosmos/gaia/v24/tests/e2e/common"
-	"github.com/cosmos/gaia/v24/tests/e2e/query"
+	"github.com/cosmos/gaia/v25/tests/e2e/common"
+	"github.com/cosmos/gaia/v25/tests/e2e/query"
 )
 
 func (s *IntegrationTestSuite) testEvidence() {