+++ tests/interchain/delegator/liquid_test.go
@@ -3,18 +3,18 @@ package delegator_test
 import (
 	"encoding/json"
 	"fmt"
-	"path"
+	"strconv"
 	"testing"
 	"time"
 
 	sdkmath "cosmossdk.io/math"
-	"github.com/cosmos/gaia/v24/tests/interchain/chainsuite"
+	"github.com/cosmos/gaia/v25/tests/interchain/chainsuite"
 	"github.com/strangelove-ventures/interchaintest/v8"
 	"github.com/strangelove-ventures/interchaintest/v8/ibc"
 	"github.com/strangelove-ventures/interchaintest/v8/testutil"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
-	"golang.org/x/mod/semver"
+	"github.com/tidwall/gjson"
+	"github.com/tidwall/sjson"
 	"golang.org/x/sync/errgroup"
 )
 
@@ -28,10 +28,17 @@ const (
 
 type LSMSuite struct {
 	*chainsuite.Suite
-	Stride      *chainsuite.Chain
-	ICAAddr     string
+	LinkedChain *chainsuite.Chain
 	LSMWallets  map[string]ibc.Wallet
 	ShareFactor sdkmath.Int
+}
+
+type ProposalJSON struct {
+	Messages       []json.RawMessage `json:"messages"`
+	InitialDeposit string            `json:"deposit"`
+	Title          string            `json:"title"`
+	Summary        string            `json:"summary"`
+	Metadata       string            `json:"metadata"`
 }
 
 func (s *LSMSuite) checkAMinusBEqualsX(a, b string, x sdkmath.Int) {
@@ -51,29 +58,7 @@ func (s *LSMSuite) TestLSMHappyPath() {
 		liquid1Redeem = 20000000
 	)
 	providerWallet := s.Chain.ValidatorWallets[0]
-
-	strideWallet := s.Stride.ValidatorWallets[0]
-
-	s.Run("Validator Bond", func() {
-		delegatorShares1, err := s.Chain.QueryJSON(s.GetContext(), "validator.delegator_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		validatorBondShares1, err := s.Chain.QueryJSON(s.GetContext(), "validator.validator_bond_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmBondingMoniker].FormattedAddress(),
-			"staking", "delegate", providerWallet.ValoperAddress, fmt.Sprintf("%d%s", delegation, s.Chain.Config().Denom))
-		s.Require().NoError(err)
-		delegatorShares2, err := s.Chain.QueryJSON(s.GetContext(), "validator.delegator_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		s.checkAMinusBEqualsX(delegatorShares2.String(), delegatorShares1.String(), sdkmath.NewInt(delegation))
-
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmBondingMoniker].FormattedAddress(),
-			"staking", "validator-bond", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		validatorBondShares2, err := s.Chain.QueryJSON(s.GetContext(), "validator.validator_bond_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		s.checkAMinusBEqualsX(validatorBondShares2.String(), validatorBondShares1.String(), sdkmath.NewInt(delegation).Mul(s.ShareFactor))
-	})
+	ibcWallet := s.LinkedChain.ValidatorWallets[0]
 
 	var tokenizedDenom string
 	s.Run("Tokenize", func() {
@@ -86,14 +71,17 @@ func (s *LSMSuite) TestLSMHappyPath() {
 		s.Require().NoError(err)
 		s.checkAMinusBEqualsX(delegatorShares2.String(), delegatorShares1.String(), sdkmath.NewInt(delegation))
 
-		sharesPreTokenize, err := s.Chain.QueryJSON(s.GetContext(), "validator.liquid_shares", "staking", "validator", providerWallet.ValoperAddress)
+		sharesPreTokenize, err := s.Chain.QueryJSON(s.GetContext(), "liquid_validator.liquid_shares", "liquid",
+			"liquid-validator",
+			providerWallet.ValoperAddress)
 		s.Require().NoError(err)
 		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmLiquid1Moniker].FormattedAddress(),
-			"staking", "tokenize-share",
+			"liquid", "tokenize-share",
 			providerWallet.ValoperAddress, fmt.Sprintf("%d%s", tokenize, s.Chain.Config().Denom), s.LSMWallets[lsmLiquid1Moniker].FormattedAddress(),
 			"--gas", "auto")
 		s.Require().NoError(err)
-		sharesPostTokenize, err := s.Chain.QueryJSON(s.GetContext(), "validator.liquid_shares", "staking", "validator", providerWallet.ValoperAddress)
+		sharesPostTokenize, err := s.Chain.QueryJSON(s.GetContext(), "liquid_validator.liquid_shares", "liquid",
+			"liquid-validator", providerWallet.ValoperAddress)
 		s.Require().NoError(err)
 		s.checkAMinusBEqualsX(sharesPostTokenize.String(), sharesPreTokenize.String(), sdkmath.NewInt(tokenize).Mul(s.ShareFactor))
 
@@ -108,28 +96,32 @@ func (s *LSMSuite) TestLSMHappyPath() {
 	})
 
 	s.Run("Transfer Ownership", func() {
-		recordIDResult, err := s.Chain.QueryJSON(s.GetContext(), "record.id", "staking", "tokenize-share-record-by-denom", tokenizedDenom)
+		recordIDResult, err := s.Chain.QueryJSON(s.GetContext(), "record.id", "liquid",
+			"tokenize-share-record-by-denom", tokenizedDenom)
 		s.Require().NoError(err)
 		recordID := recordIDResult.String()
 
-		ownerResult, err := s.Chain.QueryJSON(s.GetContext(), "record.owner", "staking", "tokenize-share-record-by-denom", tokenizedDenom)
+		ownerResult, err := s.Chain.QueryJSON(s.GetContext(), "record.owner", "liquid",
+			"tokenize-share-record-by-denom", tokenizedDenom)
 		s.Require().NoError(err)
 		owner := ownerResult.String()
 
 		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), owner,
-			"staking", "transfer-tokenize-share-record", recordID, s.LSMWallets[lsmOwnerMoniker].FormattedAddress())
+			"liquid", "transfer-tokenize-share-record", recordID, s.LSMWallets[lsmOwnerMoniker].FormattedAddress())
 		s.Require().NoError(err)
 
-		ownerResult, err = s.Chain.QueryJSON(s.GetContext(), "record.owner", "staking", "tokenize-share-record-by-denom", tokenizedDenom)
+		ownerResult, err = s.Chain.QueryJSON(s.GetContext(), "record.owner", "liquid",
+			"tokenize-share-record-by-denom", tokenizedDenom)
 		s.Require().NoError(err)
 		owner = ownerResult.String()
 		s.Require().Equal(s.LSMWallets[lsmOwnerMoniker].FormattedAddress(), owner)
 
 		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), owner,
-			"staking", "transfer-tokenize-share-record", recordID, s.LSMWallets[lsmLiquid1Moniker].FormattedAddress())
+			"liquid", "transfer-tokenize-share-record", recordID, s.LSMWallets[lsmLiquid1Moniker].FormattedAddress())
 		s.Require().NoError(err)
 
-		ownerResult, err = s.Chain.QueryJSON(s.GetContext(), "record.owner", "staking", "tokenize-share-record-by-denom", tokenizedDenom)
+		ownerResult, err = s.Chain.QueryJSON(s.GetContext(), "record.owner", "liquid",
+			"tokenize-share-record-by-denom", tokenizedDenom)
 		s.Require().NoError(err)
 		owner = ownerResult.String()
 		s.Require().Equal(s.LSMWallets[lsmLiquid1Moniker].FormattedAddress(), owner)
@@ -138,9 +130,9 @@ func (s *LSMSuite) TestLSMHappyPath() {
 	var happyLiquid1Delegations1 string
 	var ibcDenom string
 
-	ibcChannelProvider, err := s.Relayer.GetTransferChannel(s.GetContext(), s.Chain, s.Stride)
+	ibcChannelProvider, err := s.Relayer.GetTransferChannel(s.GetContext(), s.Chain, s.LinkedChain)
 	s.Require().NoError(err)
-	ibcChannelStride, err := s.Relayer.GetTransferChannel(s.GetContext(), s.Stride, s.Chain)
+	ibcChannel, err := s.Relayer.GetTransferChannel(s.GetContext(), s.LinkedChain, s.Chain)
 	s.Require().NoError(err)
 
 	s.Run("Transfer Tokens", func() {
@@ -158,11 +150,11 @@ func (s *LSMSuite) TestLSMHappyPath() {
 		_, err = s.Chain.SendIBCTransfer(s.GetContext(), ibcChannelProvider.ChannelID, s.LSMWallets[lsmLiquid1Moniker].FormattedAddress(), ibc.WalletAmount{
 			Amount:  sdkmath.NewInt(ibcTransfer),
 			Denom:   tokenizedDenom,
-			Address: strideWallet.Address,
+			Address: ibcWallet.Address,
 		}, ibc.TransferOptions{})
 		s.Require().NoError(err)
-		s.Require().NoError(testutil.WaitForBlocks(s.GetContext(), 5, s.Stride))
-		balances, err := s.Stride.BankQueryAllBalances(s.GetContext(), strideWallet.Address)
+		s.Require().NoError(testutil.WaitForBlocks(s.GetContext(), 5, s.LinkedChain))
+		balances, err := s.LinkedChain.BankQueryAllBalances(s.GetContext(), ibcWallet.Address)
 		s.Require().NoError(err)
 		for _, balance := range balances {
 			if balance.Amount.Int64() == ibcTransfer {
@@ -175,16 +167,16 @@ func (s *LSMSuite) TestLSMHappyPath() {
 	var happyLiquid1DelegationBalance string
 	s.Run("Redeem Tokens", func() {
 		_, err := s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmLiquid1Moniker].FormattedAddress(),
-			"staking", "redeem-tokens", fmt.Sprintf("%d%s", liquid1Redeem, tokenizedDenom),
+			"liquid", "redeem-tokens", fmt.Sprintf("%d%s", liquid1Redeem, tokenizedDenom),
 			"--gas", "auto")
 		s.Require().NoError(err)
 
 		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmLiquid2Moniker].FormattedAddress(),
-			"staking", "redeem-tokens", fmt.Sprintf("%d%s", bankSend, tokenizedDenom),
+			"liquid", "redeem-tokens", fmt.Sprintf("%d%s", bankSend, tokenizedDenom),
 			"--gas", "auto")
 		s.Require().NoError(err)
 
-		_, err = s.Stride.SendIBCTransfer(s.GetContext(), ibcChannelStride.ChannelID, strideWallet.Address, ibc.WalletAmount{
+		_, err = s.LinkedChain.SendIBCTransfer(s.GetContext(), ibcChannel.ChannelID, ibcWallet.Address, ibc.WalletAmount{
 			Amount:  sdkmath.NewInt(ibcTransfer),
 			Denom:   ibcDenom,
 			Address: s.LSMWallets[lsmLiquid3Moniker].FormattedAddress(),
@@ -194,7 +186,7 @@ func (s *LSMSuite) TestLSMHappyPath() {
 		s.Require().NoError(testutil.WaitForBlocks(s.GetContext(), 5, s.Chain))
 
 		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmLiquid3Moniker].FormattedAddress(),
-			"staking", "redeem-tokens", fmt.Sprintf("%d%s", ibcTransfer, tokenizedDenom),
+			"liquid", "redeem-tokens", fmt.Sprintf("%d%s", ibcTransfer, tokenizedDenom),
 			"--gas", "auto")
 		s.Require().NoError(err)
 
@@ -229,154 +221,6 @@ func (s *LSMSuite) TestLSMHappyPath() {
 		s.checkAMinusBEqualsX(happyLiquid1DelegationBalance, "0", sdkmath.NewInt(70000000))
 		s.checkAMinusBEqualsX(happyLiquid2DelegationBalance, "0", sdkmath.NewInt(bankSend))
 		s.checkAMinusBEqualsX(happyLiquid3DelegationBalance, "0", sdkmath.NewInt(ibcTransfer))
-	})
-	s.Run("Cleanup", func() {
-		validatorBondSharesBeforeResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.validator_bond_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		validatorBondSharesBefore := validatorBondSharesBeforeResult.String()
-
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmBondingMoniker].FormattedAddress(),
-			"staking", "unbond", providerWallet.ValoperAddress, fmt.Sprintf("%d%s", delegation, s.Chain.Config().Denom))
-		s.Require().NoError(err)
-
-		validatorBondSharesResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.validator_bond_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		validatorBondShares := validatorBondSharesResult.String()
-		s.checkAMinusBEqualsX(validatorBondSharesBefore, validatorBondShares, sdkmath.NewInt(delegation).Mul(s.ShareFactor))
-
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmLiquid1Moniker].FormattedAddress(),
-			"staking", "unbond", providerWallet.ValoperAddress, fmt.Sprintf("%s%s", happyLiquid1DelegationBalance, s.Chain.Config().Denom))
-		s.Require().NoError(err)
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmLiquid2Moniker].FormattedAddress(),
-			"staking", "unbond", providerWallet.ValoperAddress, fmt.Sprintf("%d%s", bankSend, s.Chain.Config().Denom))
-		s.Require().NoError(err)
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.LSMWallets[lsmLiquid3Moniker].FormattedAddress(),
-			"staking", "unbond", providerWallet.ValoperAddress, fmt.Sprintf("%d%s", ibcTransfer, s.Chain.Config().Denom))
-		s.Require().NoError(err)
-	})
-}
-
-func (s *LSMSuite) TestICADelegate() {
-	const (
-		delegate       = 20000000
-		bondDelegation = 20000000
-	)
-	bondingWallet, err := s.Chain.BuildWallet(s.GetContext(), fmt.Sprintf("lsm_happy_bonding_%d", time.Now().Unix()), "")
-	s.Require().NoError(err)
-
-	err = s.Chain.SendFunds(s.GetContext(), interchaintest.FaucetAccountKeyName, ibc.WalletAmount{
-		Amount:  sdkmath.NewInt(50_000_000),
-		Denom:   s.Chain.Config().Denom,
-		Address: bondingWallet.FormattedAddress(),
-	})
-	s.Require().NoError(err)
-
-	providerWallet := s.Chain.ValidatorWallets[0]
-
-	strideWallet := s.Stride.ValidatorWallets[0]
-
-	s.Run("Delegate and Bond", func() {
-		shares1Result, err := s.Chain.QueryJSON(s.GetContext(), "validator.delegator_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		shares1 := shares1Result.String()
-
-		tokens1Result, err := s.Chain.QueryJSON(s.GetContext(), "validator.tokens", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		tokens1 := tokens1Result.String()
-
-		bondShares1Result, err := s.Chain.QueryJSON(s.GetContext(), "validator.validator_bond_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		bondShares1 := bondShares1Result.String()
-
-		shares1Int, err := chainsuite.StrToSDKInt(shares1)
-		s.Require().NoError(err)
-		tokens1Int, err := chainsuite.StrToSDKInt(tokens1)
-		s.Require().NoError(err)
-		bondShares1Int, err := chainsuite.StrToSDKInt(bondShares1)
-		s.Require().NoError(err)
-
-		exchangeRate1 := shares1Int.Quo(tokens1Int)
-		expectedSharesIncrease := exchangeRate1.MulRaw(bondDelegation).Mul(s.ShareFactor)
-		expectedShares := expectedSharesIncrease.Add(bondShares1Int)
-
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), bondingWallet.FormattedAddress(),
-			"staking", "delegate", providerWallet.ValoperAddress, fmt.Sprintf("%d%s", bondDelegation, s.Chain.Config().Denom))
-		s.Require().NoError(err)
-
-		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), bondingWallet.FormattedAddress(),
-			"staking", "validator-bond", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-
-		bondShares2Result, err := s.Chain.QueryJSON(s.GetContext(), "validator.validator_bond_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		bondShares2 := bondShares2Result.String()
-
-		bondShares2Int, err := chainsuite.StrToSDKInt(bondShares2)
-		s.Require().NoError(err)
-		s.Require().Truef(bondShares2Int.Sub(expectedShares).Abs().LTE(sdkmath.NewInt(1)), "bondShares2: %s, expectedShares: %s", bondShares2, expectedShares)
-	})
-
-	s.Run("Delegate via ICA", func() {
-		preDelegationTokensResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.tokens", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		preDelegationTokens := preDelegationTokensResult.String()
-
-		preDelegationSharesResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.delegator_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		preDelegationShares := preDelegationSharesResult.String()
-
-		preDelegationLiquidSharesResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.liquid_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		preDelegationLiquidShares := preDelegationLiquidSharesResult.String()
-
-		preDelegationTokensInt, err := chainsuite.StrToSDKInt(preDelegationTokens)
-		s.Require().NoError(err)
-		preDelegationSharesInt, err := chainsuite.StrToSDKInt(preDelegationShares)
-		s.Require().NoError(err)
-		exchangeRate := preDelegationSharesInt.Quo(preDelegationTokensInt)
-		expectedLiquidIncrease := exchangeRate.MulRaw(delegate).Mul(s.ShareFactor)
-
-		delegateHappy := map[string]interface{}{
-			"@type":             "/cosmos.staking.v1beta1.MsgDelegate",
-			"delegator_address": s.ICAAddr,
-			"validator_address": providerWallet.ValoperAddress,
-			"amount": map[string]interface{}{
-				"denom":  s.Chain.Config().Denom,
-				"amount": fmt.Sprint(delegate),
-			},
-		}
-		delegateHappyJSON, err := json.Marshal(delegateHappy)
-		s.Require().NoError(err)
-		jsonPath := "delegate-happy.json"
-		fullJsonPath := path.Join(s.Stride.Validators[0].HomeDir(), jsonPath)
-		stdout, _, err := s.Stride.GetNode().ExecBin(s.GetContext(), "tx", "interchain-accounts", "host", "generate-packet-data", string(delegateHappyJSON), "--encoding", "proto3")
-		s.Require().NoError(err)
-		s.Require().NoError(s.Stride.Validators[0].WriteFile(s.GetContext(), []byte(stdout), jsonPath))
-		ibcChannelStride, err := s.Relayer.GetTransferChannel(s.GetContext(), s.Stride, s.Chain)
-		s.Require().NoError(err)
-
-		_, err = s.Stride.GetNode().ExecTx(s.GetContext(), strideWallet.Address,
-			"interchain-accounts", "controller", "send-tx", ibcChannelStride.ConnectionHops[0], fullJsonPath)
-		s.Require().NoError(err)
-
-		var tokensDelta sdkmath.Int
-		s.Require().EventuallyWithT(func(c *assert.CollectT) {
-			postDelegationTokensResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.tokens", "staking", "validator", providerWallet.ValoperAddress)
-			s.Require().NoError(err)
-			postDelegationTokens, err := chainsuite.StrToSDKInt(postDelegationTokensResult.String())
-			s.Require().NoError(err)
-			tokensDelta = postDelegationTokens.Sub(preDelegationTokensInt)
-			assert.Truef(c, tokensDelta.Sub(sdkmath.NewInt(delegate)).Abs().LTE(sdkmath.NewInt(1)), "tokensDelta: %s, delegate: %d", tokensDelta, delegate)
-		}, 20*time.Second, time.Second)
-
-		postDelegationLiquidSharesResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.liquid_shares", "staking", "validator", providerWallet.ValoperAddress)
-		s.Require().NoError(err)
-		postDelegationLiquidShares, err := chainsuite.StrToSDKInt(postDelegationLiquidSharesResult.String())
-		s.Require().NoError(err)
-		preDelegationLiquidSharesInt, err := chainsuite.StrToSDKInt(preDelegationLiquidShares)
-		s.Require().NoError(err)
-		liquidSharesDelta := postDelegationLiquidShares.Sub(preDelegationLiquidSharesInt)
-		s.Require().Truef(liquidSharesDelta.Sub(expectedLiquidIncrease).Abs().LTE(sdkmath.NewInt(1)), "liquidSharesDelta: %s, expectedLiquidIncrease: %s", liquidSharesDelta, expectedLiquidIncrease)
 	})
 }
 
@@ -413,24 +257,270 @@ func (s *LSMSuite) TestTokenizeVested() {
 
 	// try to tokenize full amount. Should fail.
 	_, err = s.Chain.GetNode().ExecTx(s.GetContext(), vestingAccount.FormattedAddress(),
-		"staking", "tokenize-share", validatorWallet.ValoperAddress, fmt.Sprintf("%d%s", vestingAmount, s.Chain.Config().Denom), vestingAccount.FormattedAddress(),
+		"liquid", "tokenize-share", validatorWallet.ValoperAddress, fmt.Sprintf("%d%s", vestingAmount,
+			s.Chain.Config().Denom), vestingAccount.FormattedAddress(),
 		"--gas", "auto")
 	s.Require().Error(err)
 
-	sharesPreTokenizeResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.liquid_shares", "staking", "validator", validatorWallet.ValoperAddress)
+	sharesPreTokenizeResult, err := s.Chain.QueryJSON(s.GetContext(), "liquid_validator.liquid_shares", "liquid",
+		"liquid-validator", validatorWallet.ValoperAddress)
 	s.Require().NoError(err)
 	sharesPreTokenize := sharesPreTokenizeResult.String()
 
 	// try to tokenize vested amount (i.e. half) should succeed if upgraded
 	tokenizeAmount := vestingAmount / 2
 	_, err = s.Chain.GetNode().ExecTx(s.GetContext(), vestingAccount.FormattedAddress(),
-		"staking", "tokenize-share", validatorWallet.ValoperAddress, fmt.Sprintf("%d%s", tokenizeAmount, s.Chain.Config().Denom), vestingAccount.FormattedAddress(),
+		"liquid", "tokenize-share", validatorWallet.ValoperAddress, fmt.Sprintf("%d%s", tokenizeAmount,
+			s.Chain.Config().Denom), vestingAccount.FormattedAddress(),
 		"--gas", "auto")
 	s.Require().NoError(err)
-	sharesPostTokenizeResult, err := s.Chain.QueryJSON(s.GetContext(), "validator.liquid_shares", "staking", "validator", validatorWallet.ValoperAddress)
+	sharesPostTokenizeResult, err := s.Chain.QueryJSON(s.GetContext(), "liquid_validator.liquid_shares", "liquid",
+		"liquid-validator", validatorWallet.ValoperAddress)
 	s.Require().NoError(err)
 	sharesPostTokenize := sharesPostTokenizeResult.String()
 	s.checkAMinusBEqualsX(sharesPostTokenize, sharesPreTokenize, sdkmath.NewInt(tokenizeAmount).Mul(s.ShareFactor))
+}
+
+func (s *LSMSuite) TestLSMParams() {
+	const (
+		delegation   = 100000000
+		globalCap    = 0.1
+		validatorCap = 0.05
+	)
+
+	providerWallet := s.Chain.ValidatorWallets[0]
+
+	liquidParams, err := s.Chain.QueryJSON(s.GetContext(), "params", "liquid", "params")
+	s.Require().NoError(err)
+
+	startingValidatorCap := gjson.Get(liquidParams.String(), "validator_liquid_staking_cap").String()
+	startingGlobalCap := gjson.Get(liquidParams.String(), "global_liquid_staking_cap").String()
+	chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Starting validator liquid staking cap: %s", startingValidatorCap)
+	chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Starting global liquid staking cap: %s", startingGlobalCap)
+
+	s.Run("Update params", func() {
+		authority, err := s.Chain.GetGovernanceAddress(s.GetContext())
+		s.Require().NoError(err)
+
+		updatedParams, err := sjson.Set(liquidParams.String(), "global_liquid_staking_cap", fmt.Sprintf("%f", globalCap))
+		s.Require().NoError(err)
+		updatedParams, err = sjson.Set(updatedParams, "validator_liquid_staking_cap", fmt.Sprintf("%f", validatorCap))
+		s.Require().NoError(err)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Updated params: %s", updatedParams)
+
+		paramChangeMessage := fmt.Sprintf(`{
+		"@type": "/gaia.liquid.v1beta1.MsgUpdateParams",
+		"authority": "%s",
+		"params": %s
+	}`, authority, updatedParams)
+
+		proposal := ProposalJSON{
+			Messages:       []json.RawMessage{json.RawMessage(paramChangeMessage)},
+			InitialDeposit: "5000000uatom",
+			Title:          "Liquid Param Change Proposal",
+			Summary:        "Test Proposal",
+			Metadata:       "ipfs://CID",
+		}
+
+		// Marshal to JSON
+		proposalBytes, err := json.MarshalIndent(proposal, "", "  ")
+		s.Require().NoError(err)
+
+		// Get the home directory of the node
+		homeDir := s.Chain.GetNode().HomeDir()
+		proposalPath := homeDir + "/params-proposal.json"
+		// Write to file
+		err = s.Chain.GetNode().WriteFile(s.GetContext(), proposalBytes, "params-proposal.json")
+		s.Require().NoError(err)
+
+		// out, _, _ := s.Chain.GetNode().Exec(s.GetContext(), []string{"ls", "-l", string(homeDir)}, []string{homeDir})
+		// fmt.Println("Files in home dir:", string(out))
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), providerWallet.Address,
+			"gov", "submit-proposal", proposalPath)
+		s.Require().NoError(err)
+
+		lastProposalId, err := s.Chain.QueryJSON(s.GetContext(), "proposals.@reverse.0.id", "gov", "proposals")
+		s.Require().NoError(err)
+
+		// Pass proposal
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), providerWallet.Address,
+			"gov", "vote", lastProposalId.String(), "yes")
+		s.Require().NoError(err)
+		time.Sleep(80 * time.Second)
+
+		// Check proposal
+		proposalStatus, err := s.Chain.QueryJSON(s.GetContext(), "proposal", "gov", "proposal", lastProposalId.String())
+		s.Require().NoError(err)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Proposal status: %s", proposalStatus)
+		// Check params
+		liquidParams, err = s.Chain.QueryJSON(s.GetContext(), "params", "liquid", "params")
+		s.Require().NoError(err)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Current params: %s", liquidParams)
+		validatorLiquidCapParam := gjson.Get(liquidParams.String(), "validator_liquid_staking_cap")
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Validator liquid cap: %s", validatorLiquidCapParam)
+		globalLiquidCapParam := gjson.Get(liquidParams.String(), "global_liquid_staking_cap")
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Global liquid cap: %s", globalLiquidCapParam)
+		validatorLiquidCapFloat := validatorLiquidCapParam.Float()
+		globalLiquidCapFloat := globalLiquidCapParam.Float()
+		s.Require().Equal(validatorCap, validatorLiquidCapFloat)
+		s.Require().Equal(globalCap, globalLiquidCapFloat)
+	})
+
+	s.Run("Test liquid caps", func() {
+		_, err := s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "staking", "delegate", s.Chain.ValidatorWallets[1].ValoperAddress, fmt.Sprintf("%d%s", delegation, chainsuite.Uatom))
+		s.Require().NoError(err)
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "staking", "delegate", s.Chain.ValidatorWallets[2].ValoperAddress, fmt.Sprintf("%d%s", delegation, chainsuite.Uatom))
+		s.Require().NoError(err)
+
+		liquidParams, err := s.Chain.QueryJSON(s.GetContext(), "params", "liquid", "params")
+		s.Require().NoError(err)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Starting params: %s", liquidParams)
+		validatorLiquidCapParam := gjson.Get(liquidParams.String(), "validator_liquid_staking_cap")
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Validator liquid cap: %s", validatorLiquidCapParam)
+		globalLiquidCapParam := gjson.Get(liquidParams.String(), "global_liquid_staking_cap")
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Global liquid cap: %s", globalLiquidCapParam)
+
+		val0Tokens, err := s.Chain.QueryJSON(s.GetContext(), "validator.tokens", "staking", "validator", s.Chain.ValidatorWallets[0].ValoperAddress)
+		s.Require().NoError(err)
+		val1Tokens, err := s.Chain.QueryJSON(s.GetContext(), "validator.tokens", "staking", "validator", s.Chain.ValidatorWallets[1].ValoperAddress)
+		s.Require().NoError(err)
+		val2Tokens, err := s.Chain.QueryJSON(s.GetContext(), "validator.tokens", "staking", "validator", s.Chain.ValidatorWallets[2].ValoperAddress)
+		s.Require().NoError(err)
+		totalBondedTokens, err := s.Chain.QueryJSON(s.GetContext(), "pool.bonded_tokens", "staking", "pool")
+		s.Require().NoError(err)
+		totalLiquidStaked, err := s.Chain.QueryJSON(s.GetContext(), "tokens", "liquid", "total-liquid-staked")
+		s.Require().NoError(err)
+		val0TokensFloat := val0Tokens.Float()
+		val1TokensFloat := val1Tokens.Float()
+		val2TokensFloat := val2Tokens.Float()
+		totalBondedTokensFloat := totalBondedTokens.Float()
+		totalLiquidStakedFloat := totalLiquidStaked.Float()
+
+		validatorCap := validatorLiquidCapParam.Float()
+		globalCap := globalLiquidCapParam.Float()
+
+		val0Cap := validatorCap * val0TokensFloat
+		val1Cap := validatorCap * val1TokensFloat
+		val2Cap := validatorCap * val2TokensFloat
+		globalCapShares := globalCap * totalBondedTokensFloat
+
+		val0LiquidShares, err := s.Chain.QueryJSON(s.GetContext(), "liquid_validator.liquid_shares", "liquid", "liquid-validator", s.Chain.ValidatorWallets[0].ValoperAddress)
+		s.Require().NoError(err)
+		val1LiquidShares, err := s.Chain.QueryJSON(s.GetContext(), "liquid_validator.liquid_shares", "liquid", "liquid-validator", s.Chain.ValidatorWallets[1].ValoperAddress)
+		s.Require().NoError(err)
+		val2LiquidShares, err := s.Chain.QueryJSON(s.GetContext(), "liquid_validator.liquid_shares", "liquid", "liquid-validator", s.Chain.ValidatorWallets[2].ValoperAddress)
+		s.Require().NoError(err)
+		val0LiquidSharesFloat := val0LiquidShares.Float()
+		val1LiquidSharesFloat := val1LiquidShares.Float()
+		val2LiquidSharesFloat := val2LiquidShares.Float()
+
+		val0CapAvailable := val0Cap - val0LiquidSharesFloat
+		val1CapAvailable := val1Cap - val1LiquidSharesFloat
+		val2CapAvailable := val2Cap - val2LiquidSharesFloat
+		globalCapSharesAvailable := globalCapShares - totalLiquidStakedFloat
+
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("val0 cap: %f, available shares: %f", val0Cap, val0CapAvailable)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("val1 cap: %f, available shares: %f", val1Cap, val1CapAvailable)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Val2 cap: %f, available shares: %f", val2Cap, val2CapAvailable)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Global cap: %f, available shares: %f", globalCapShares, globalCapSharesAvailable)
+
+		// Validator 1 must have a lower cap than the global amount
+		s.Require().Less(val1CapAvailable, globalCapSharesAvailable)
+
+		// Try to tokenize more than the global cap
+		globalFailAmount := int(globalCapSharesAvailable + 1000000)
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "liquid", "tokenize-share",
+			s.Chain.ValidatorWallets[1].ValoperAddress, fmt.Sprintf("%d%s", globalFailAmount, s.Chain.Config().Denom), s.Chain.ValidatorWallets[0].Address)
+		s.Require().Error(err)
+
+		// Try to tokenize more than the validator cap
+		validatorFailAmount := int(val1CapAvailable + 1000000)
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "liquid", "tokenize-share",
+			s.Chain.ValidatorWallets[1].ValoperAddress, fmt.Sprintf("%d%s", validatorFailAmount, s.Chain.Config().Denom), s.Chain.ValidatorWallets[0].Address)
+		s.Require().Error(err)
+
+		// Tokenize less than the validator cap
+		validatorSuccessAmount := int(val1CapAvailable - 1000000)
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "liquid", "tokenize-share",
+			s.Chain.ValidatorWallets[1].ValoperAddress, fmt.Sprintf("%d%s", validatorSuccessAmount, s.Chain.Config().Denom), s.Chain.ValidatorWallets[0].Address)
+		s.Require().NoError(err)
+
+		tokenizedDenom, err := s.Chain.QueryJSON(s.GetContext(), "balances.@reverse.1.denom", "bank", "balances", s.Chain.ValidatorWallets[0].Address)
+		s.Require().NoError(err)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Tokenized denom: %s", tokenizedDenom)
+		// Redeem tokenized amount
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "liquid", "redeem-tokens",
+			fmt.Sprintf("%d%s", validatorSuccessAmount, tokenizedDenom))
+		s.Require().NoError(err)
+
+		// Unbond
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "staking", "unbond", s.Chain.ValidatorWallets[1].ValoperAddress, fmt.Sprintf("%d%s", delegation, chainsuite.Uatom))
+		s.Require().NoError(err)
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), s.Chain.ValidatorWallets[0].Address, "staking", "unbond", s.Chain.ValidatorWallets[2].ValoperAddress, fmt.Sprintf("%d%s", delegation, chainsuite.Uatom))
+		s.Require().NoError(err)
+	})
+
+	s.Run("Restore params", func() {
+		authority, err := s.Chain.GetGovernanceAddress(s.GetContext())
+		s.Require().NoError(err)
+
+		updatedParams, err := sjson.Set(liquidParams.String(), "global_liquid_staking_cap", startingGlobalCap)
+		s.Require().NoError(err)
+		updatedParams, err = sjson.Set(updatedParams, "validator_liquid_staking_cap", startingValidatorCap)
+		s.Require().NoError(err)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Updated params: %s", updatedParams)
+
+		paramChangeMessage := fmt.Sprintf(`{
+				"@type": "/gaia.liquid.v1beta1.MsgUpdateParams",
+				"authority": "%s",
+				"params": %s
+			}`, authority, updatedParams)
+
+		proposal := ProposalJSON{
+			Messages:       []json.RawMessage{json.RawMessage(paramChangeMessage)},
+			InitialDeposit: "5000000uatom",
+			Title:          "Liquid Param Restore Proposal",
+			Summary:        "Test Proposal",
+			Metadata:       "ipfs://CID",
+		}
+
+		// Marshal to JSON
+		proposalBytes, err := json.MarshalIndent(proposal, "", "  ")
+		s.Require().NoError(err)
+
+		// Get the home directory of the node
+		homeDir := s.Chain.GetNode().HomeDir()
+		proposalPath := homeDir + "/restore-params-proposal.json"
+		// Write to file
+		err = s.Chain.GetNode().WriteFile(s.GetContext(), proposalBytes, "restore-params-proposal.json")
+		s.Require().NoError(err)
+
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), providerWallet.Address,
+			"gov", "submit-proposal", proposalPath)
+		s.Require().NoError(err)
+
+		lastProposalId, err := s.Chain.QueryJSON(s.GetContext(), "proposals.@reverse.0.id", "gov", "proposals")
+		s.Require().NoError(err)
+		chainsuite.GetLogger(s.GetContext()).Sugar().Infof("Last Proposal ID: %s", lastProposalId)
+
+		// Pass proposal
+		_, err = s.Chain.GetNode().ExecTx(s.GetContext(), providerWallet.Address,
+			"gov", "vote", lastProposalId.String(), "yes")
+		s.Require().NoError(err)
+		time.Sleep(80 * time.Second)
+
+		// Check params
+		liquidParams, err = s.Chain.QueryJSON(s.GetContext(), "params", "liquid", "params")
+		s.Require().NoError(err)
+		validatorLiquidCapParam := gjson.Get(liquidParams.String(), "validator_liquid_staking_cap").Float()
+		globalLiquidCapParam := gjson.Get(liquidParams.String(), "global_liquid_staking_cap").Float()
+		startingValidatorCapFloat, err := strconv.ParseFloat(startingValidatorCap, 64)
+		s.Require().NoError(err)
+		startingGlobalCapFloat, err := strconv.ParseFloat(startingGlobalCap, 64)
+		s.Require().NoError(err)
+		s.Require().Equal(startingValidatorCapFloat, validatorLiquidCapParam)
+		s.Require().Equal(startingGlobalCapFloat, globalLiquidCapParam)
+	})
 }
 
 func (s *LSMSuite) setupLSMWallets() {
@@ -460,38 +550,25 @@ func (s *LSMSuite) setupLSMWallets() {
 
 func (s *LSMSuite) SetupSuite() {
 	s.Suite.SetupSuite()
-	// This is slightly broken while stride is still in the process of being upgraded, so skip if
-	// going from v21 -> v21
-	if semver.Major(s.Env.OldGaiaImageVersion) == s.Env.UpgradeName && s.Env.UpgradeName == "v21" {
-		s.T().Skip("Skipping LSM when going from v21 -> v21")
-	}
-	stride, err := s.Chain.AddConsumerChain(s.GetContext(), s.Relayer, chainsuite.ConsumerConfig{
-		ChainName:             "stride",
-		Version:               chainsuite.StrideVersion,
-		Denom:                 chainsuite.StrideDenom,
-		TopN:                  100,
-		ShouldCopyProviderKey: []bool{true},
-	})
+	secondChain, err := s.Chain.AddLinkedChain(s.GetContext(), s.T(), s.Relayer, chainsuite.DefaultChainSpec(s.Env))
 	s.Require().NoError(err)
-	s.Stride = stride
-	err = s.Chain.CheckCCV(s.GetContext(), stride, s.Relayer, 1_000_000, 0, 1)
-	s.Require().NoError(err)
+	s.LinkedChain = secondChain
 
-	icaAddr, err := stride.SetupICAAccount(s.GetContext(), s.Chain, s.Relayer, stride.ValidatorWallets[0].Address, 0, 1_000_000_000)
-	s.Require().NoError(err)
-	s.ICAAddr = icaAddr
 	shareFactor, ok := sdkmath.NewIntFromString("1000000000000000000")
 	s.Require().True(ok)
 	s.ShareFactor = shareFactor
 
 	s.setupLSMWallets()
-	s.UpgradeChain()
 }
 
 func TestLSM(t *testing.T) {
 	s := &LSMSuite{
 		Suite: chainsuite.NewSuite(chainsuite.SuiteConfig{
-			CreateRelayer: true,
+			CreateRelayer:  true,
+			UpgradeOnSetup: true,
+			ChainSpec: &interchaintest.ChainSpec{
+				NumValidators: &chainsuite.SixValidators,
+			},
 		}),
 	}
 	suite.Run(t, s)