+++ x/liquid/keeper/genesis.go
@@ -6,12 +6,16 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	"github.com/cosmos/gaia/v24/x/liquid/types"
+	"github.com/cosmos/gaia/v25/x/liquid/types"
 )
 
 // InitGenesis sets liquid information for genesis
 func (k Keeper) InitGenesis(ctx context.Context, data *types.GenesisState) {
-	//
+	// Set the x/liquid module parameters
+	if err := k.SetParams(ctx, data.Params); err != nil {
+		panic(err)
+	}
+
 	// Set the total liquid staked tokens
 	k.SetTotalLiquidStakedTokens(ctx, data.TotalLiquidStakedTokens)